Moving TimeFreezer into utils.py
